Refetch Data after Datepicker input
- Fixed datepicker component to re-render on state change

- [fetchAndDisplayNasaData.ts] Broke fetch and display functions into separate file
